Add a linker script
Our board has two memory areas: flash and RAM. The former is read-only, so we place .text and .rodata in it. The .data and .bss sections and the stack go into the RAM area.

The beginning of the flash area must contain the stack pointer (we set it to the end of RAM) and the reset entry point. After that, the exception and interrupt handlers are specified (we will do this later).

We need to define __DATA_START, __BSS_START, etc. symbols since we need to initialize these sections later in our Rust code.