We want our custom `reset` entry point, so we use #![no_main]
With the `no_main` attribute, we no longer need a `main` function. However, we still keep it for familiarity (with a small modification: main is now divering). The real entry point is our reset function (because of the `ENTRY(reset)` line in our linker script). The function needs to be public and no_mangle so that the linker can find it. We can't return from it reasonably, so we mark the function as diverging.

The reset entry point is unsafe, since it should be only called directly by the hardware.