Add a gdb.sh that launches gdb with the correct arguments
We pass `add-auto-load-safe-path .` to enable the automatic loading of our `.gdbinit` with the `reset` and `reset-load` functions. We also connect to our board using `tar ext :4242` which is short for `target extern :4242` (port 4242 is used by st-util by default). After that, we execute our new `load-reset` function to load our executable to the board and reset it afterwards.

Now we should be able to run our executable on the board. By setting breakpoints, we can verify that everything is working correctly.