Remove semi hosting output at the start of main
Semi-hosting can be great for debugging, but it causes a lot of problems (e.g. the python script does not work on newer gdb versions, it messes up breakpoints in gdb, it halts the controller if no debugger is attached, etc.). Therefore it does not make sense to force semihosting use for an useless 'Entering main' message.