Mark `EXCEPTIONS` and `INTERRUPTS` as `EXTERN` in the linker script
This is required because otherwise the linker finishes as soon as all unresolved symbols are found. If it stumbles upon one of them by accident (because it looked at the file to find other symbols), the symbol is kept. If not, it is discarded. The `EXTERN` declaration tells the linker that it should look for the passed symbol.

This also explains why our workaround worked: If the EXCEPTIONS static is alone in a separate module, it wasn't found because the linker didn't look at the module file to find other symbols.

For more details, see https://github.com/rust-lang/rust/issues/40289#issuecomment-385609104